requirements.txt
streamlit>=1.37
web3>=6.10
python-dotenv>=1.0
pandas>=2.2
pydantic>=2.7
python-docx>=0.8.11
reportlab>=4.4
openpyxl>=3.1
PyPDF2>=3.0
fastapi>=0.111
uvicorn[standard]>=0.30
openai>=1.35
XlsxWriter>=3.1
